+++ CHANGELOG.md
@@ -1,3 +1,5 @@
 ## 1.0.0
 
-- Initial version.
+- Generate Json classes
+- Generate Hive classes
+- Adding JSON files to check for existing hive fields / classes